requirements.txt
streamlit
anthropic
pypdf
python-docx
fastapi
uvicorn[standard]
python-multipart
